+++ app/contents/components/filterBar/filterBarItems/sliderFilter.tsx
